Incsrch control loop modified for debug
In brief: when plugin is in the search loop after pressing ^I,
given that ^I hotkey macro is used, it enters an input read
loop in the control.c CollectEvents(). The combination ^Enter
shall search next occurence and ^#Enter shall search previous.
Starting at some point it does not happen in far2l anymore.

The apiEditorControl(ECTL_READINPUT, ...) does not properly
handle this key combination. The loop is modified and instrumented
to collect incoming events. We have: key down for Ctrl, then
we expect VK_RETURN with Ctrl down state, but it cames with
bKeyDown=false, as if it were a release of the Enter key. The
actual keypress with bKeyDown=true never comes to ECTL_READINPUT
if Ctrl is pressed.

This is the regression in WinPort, it used to work and still works
on Windows.

The sole purpose of this branch is to facilitate debug. It bypasses
exit from the logic if no key is pressed and writes incsrch.log in
the current far2l cwd.